Sorting debugging and further testing. Bug(s) remaining
Some minor bugs found and resolved in the sorting algorithm. Program tested all the way through, still works the same as before when sorting flag is turned off.

Current bug: since there are now two nodes on the leading edge, the influence of the leading edge points on all points aft of the LE is computed twice. Can maybe use nNode count before and after sorting algorithm to prevent this.