Sorter not fully implemented, record keeping commt only
This commit is just to catalogue what was an in progress debugging effort, but one that was not completed. Description of suspected bug(s) are below.

Bugs still remain in implementing the sorting algorithm into the full supersonic solver. It appears that disparate LE and TE nodes and control points are being created, however the computations at the TE points is not working correctly for unknown reasons. I suspect it is because the TE points are now influenced by another set of LE points, which is what I tried to track down here in this commit but was unsuccessful.